# apps/api/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "LayeringStackWearRequest": {
        "additionalProperties": false,
        "properties": {
          "notes": {
            "anyOf": [
              {
                "maxLength": 2000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "rating": {
            "anyOf": [
              {
                "maximum": 5.0,
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rating"
          },
          "worn_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Worn At"
          }
        },
        "title": "LayeringStackWearRequest",
        "type": "object"
      },
      "LayeringStackWearResponse": {
        "properties": {
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "rating": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rating"
          },
          "stack_id": {
            "format": "uuid",
            "title": "Stack Id",
            "type": "string"
          },
          "worn_at": {
            "format": "date-time",
            "title": "Worn At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "stack_id",
          "worn_at",
          "rating",
          "notes"
        ],
        "title": "LayeringStackWearResponse",
        "type": "object"
      },
...
      "RecommendationWearRequest": {
        "additionalProperties": false,
        "properties": {
          "notes": {
            "anyOf": [
              {
                "maxLength": 2000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "occasion": {
            "anyOf": [
              {
                "enum": [
                  "work",
                  "casual",
                  "date",
                  "dinner",
                  "party",
                  "formal",
                  "gym",
                  "travel",
                  "other"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Occasion"
          },
          "selected_fragrance_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Selected Fragrance Id"
          },
          "setting": {
            "anyOf": [
              {
                "maxLength": 40,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Setting"
          },
          "sprays": {
            "anyOf": [
              {
                "maximum": 30.0,
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sprays"
          },
          "worn_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Worn At"
          }
        },
        "title": "RecommendationWearRequest",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "WearFeedbackRequest": {
        "additionalProperties": false,
        "properties": {
          "comments": {
            "anyOf": [
              {
                "maxLength": 2000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Comments"
          },
          "longevity": {
            "anyOf": [
              {
                "maximum": 10.0,
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Longevity"
          },
          "projection": {
            "anyOf": [
              {
                "enum": [
                  "intimate",
                  "moderate",
                  "strong"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Projection"
          },
          "rating": {
            "anyOf": [
              {
                "maximum": 5.0,
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rating"
          }
        },
        "title": "WearFeedbackRequest",
        "type": "object"
      },
      "WearFeedbackResponse": {
        "additionalProperties": false,
        "properties": {
          "comments": {
            "anyOf": [
              {
                "maxLength": 2000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Comments"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "longevity": {
            "anyOf": [
              {
                "maximum": 10.0,
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Longevity"
          },
          "projection": {
            "anyOf": [
              {
                "enum": [
                  "intimate",
                  "moderate",
                  "strong"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Projection"
          },
          "rating": {
            "anyOf": [
              {
                "maximum": 5.0,
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rating"
          },
          "wear_log_id": {
            "format": "uuid",
            "title": "Wear Log Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "wear_log_id"
        ],
        "title": "WearFeedbackResponse",
        "type": "object"
      },
      "WearLogCreateRequest": {
        "additionalProperties": false,
        "properties": {
          "collection_item_id": {
            "format": "uuid",
            "title": "Collection Item Id",
            "type": "string"
          },
          "notes": {
            "anyOf": [
              {
                "maxLength": 2000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "occasion": {
            "anyOf": [
              {
                "enum": [
                  "work",
                  "casual",
                  "date",
                  "dinner",
                  "party",
                  "formal",
                  "gym",
                  "travel",
                  "other"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Occasion"
          },
          "setting": {
            "anyOf": [
              {
                "maxLength": 40,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Setting"
          },
          "sprays": {
            "anyOf": [
              {
                "maximum": 30.0,
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sprays"
          },
          "worn_at": {
            "format": "date-time",
            "title": "Worn At",
            "type": "string"
          }
        },
        "required": [
          "collection_item_id",
          "worn_at"
        ],
        "title": "WearLogCreateRequest",
        "type": "object"
      },
      "WearLogResponse": {
        "properties": {
          "brand_name": {
            "title": "Brand Name",
            "type": "string"
          },
          "collection_item_id": {
            "format": "uuid",
            "title": "Collection Item Id",
            "type": "string"
          },
          "fragrance_id": {
            "format": "uuid",
            "title": "Fragrance Id",
            "type": "string"
          },
          "fragrance_name": {
            "title": "Fragrance Name",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "occasion": {
            "anyOf": [
              {
                "enum": [
                  "work",
                  "casual",
                  "date",
                  "dinner",
                  "party",
                  "formal",
                  "gym",
                  "travel",
                  "other"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Occasion"
          },
          "recommendation_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Recommendation Id"
          },
          "setting": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Setting"
          },
          "sprays": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sprays"
          },
          "worn_at": {
            "format": "date-time",
            "title": "Worn At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "collection_item_id",
          "fragrance_id",
          "fragrance_name",
          "brand_name",
          "worn_at"
        ],
        "title": "WearLogResponse",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "FastAPI",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/layering/stacks/{stack_id}/wears": {
      "get": {
        "operationId": "history_api_v1_layering_stacks__stack_id__wears_get",
        "parameters": [
          {
            "in": "path",
            "name": "stack_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Stack Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/LayeringStackWearResponse"
                  },
                  "title": "Response History Api V1 Layering Stacks  Stack Id  Wears Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "History",
        "tags": [
          "layering"
        ]
      },
      "post": {
        "operationId": "wear_api_v1_layering_stacks__stack_id__wears_post",
        "parameters": [
          {
            "in": "path",
            "name": "stack_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Stack Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LayeringStackWearRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LayeringStackWearResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Wear",
        "tags": [
          "layering"
        ]
      }
    },
...
    "/api/v1/recommendations/{recommendation_id}/wear": {
      "post": {
        "operationId": "wear_api_v1_recommendations__recommendation_id__wear_post",
        "parameters": [
          {
            "in": "path",
            "name": "recommendation_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Recommendation Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecommendationWearRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WearLogResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Wear",
        "tags": [
          "wear-recommendations"
        ]
      }
    },
    "/api/v1/wear-logs": {
      "get": {
        "operationId": "list_wear_logs_api_v1_wear_logs_get",
        "parameters": [
          {
            "in": "query",
            "name": "collection_item_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Collection Item Id"
            }
          },
          {
            "in": "query",
            "name": "fragrance_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Fragrance Id"
            }
          },
          {
            "in": "query",
            "name": "worn_from",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Worn From"
            }
          },
          {
            "in": "query",
            "name": "worn_to",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Worn To"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/WearLogResponse"
                  },
                  "title": "Response List Wear Logs Api V1 Wear Logs Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Wear Logs",
        "tags": [
          "wear-logs"
        ]
      },
      "post": {
        "operationId": "add_wear_log_api_v1_wear_logs_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WearLogCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WearLogResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Add Wear Log",
        "tags": [
          "wear-logs"
        ]
      }
    },
    "/api/v1/wear-logs/{wear_id}/feedback": {
      "get": {
        "operationId": "get_feedback_api_v1_wear_logs__wear_id__feedback_get",
        "parameters": [
          {
            "in": "path",
            "name": "wear_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Wear Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/WearFeedbackResponse"
                    },
                    {
                      "type": "null"
                    }
                  ],
                  "title": "Response Get Feedback Api V1 Wear Logs  Wear Id  Feedback Get"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Feedback",
        "tags": [
          "wear-logs"
        ]
      },
      "put": {
        "operationId": "put_feedback_api_v1_wear_logs__wear_id__feedback_put",
        "parameters": [
          {
            "in": "path",
            "name": "wear_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Wear Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WearFeedbackRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WearFeedbackResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Put Feedback",
        "tags": [
          "wear-logs"
        ]
      }
    },
...
  }
}